+++ apps/web/client/src/app/project/[id]/_components/right-panel/chat-tab/code-change-display/bash-code-display.tsx
