+++ apps/web/client/src/app/project/[id]/_components/right-panel/chat-tab/error.tsx
