+++ apps/web/client/src/app/project/[id]/_components/right-panel/dev-tab/file-tree-node.tsx
