+++ apps/web/client/src/app/project/[id]/_components/right-panel/index.tsx
@@ -9,13 +9,14 @@ import { cn } from '@onlook/ui/utils';
 import { observer } from 'mobx-react-lite';
 import { useTranslations } from 'next-intl';
 import { useState } from 'react';
-import { ChatTab } from './chat/chat-tab';
-import { ChatControls } from './chat/controls';
-import { ChatHistory } from './chat/history';
-import { ChatPanelDropdown } from './chat/panel-dropdown';
+import { ChatTab } from './chat-tab';
+import { ChatControls } from './chat-tab/controls';
+import { ChatHistory } from './chat-tab/history';
+import { ChatPanelDropdown } from './chat-tab/panel-dropdown';
 
 const EDIT_PANEL_WIDTHS = {
     [EditorTabValue.CHAT]: 352,
+    [EditorTabValue.DEV]: 700,
 };
 
 export const RightPanel = observer(() => {
@@ -43,24 +44,35 @@ export const RightPanel = observer(() => {
                 maxWidth={700}
             >
                 <Tabs className='h-full' onValueChange={(value) => editorEngine.state.rightPanelTab = value as EditorTabValue} value={selectedTab} >
-                    <TabsList
-                        className='flex h-12 w-full border-b justify-between items-center bg-transparent select-none px-2'
-                    >
-                        <ChatPanelDropdown>
+                    <TabsList className='flex flex-row h-12 w-full border-b items-center bg-transparent select-none px-2 justify-between'>
+                        <div className="flex flex-row items-center gap-2 ">
+                            <ChatPanelDropdown>
+                                <TabsTrigger
+                                    className="bg-transparent py-2 px-1 text-small hover:text-foreground-hover"
+                                    value={EditorTabValue.CHAT}
+                                >
+                                    <Icons.Sparkles className="mr-1.5 mb-0.5 h-4 w-4" />
+                                    {t('editor.panels.edit.tabs.chat.name')}
+                                    < Icons.ChevronDown className="ml-1 h-3 w-3 text-muted-foreground" />
+                                </TabsTrigger>
+                            </ChatPanelDropdown>
                             <TabsTrigger
                                 className="bg-transparent py-2 px-1 text-small hover:text-foreground-hover"
-                                value={EditorTabValue.CHAT}
+                                value={EditorTabValue.DEV}
                             >
-                                <Icons.Sparkles className="mr-1.5 mb-0.5 h-4 w-4" />
-                                {t('editor.panels.edit.tabs.chat.name')}
-                                < Icons.ChevronDown className="ml-1 h-3 w-3 text-muted-foreground" />
+                                <Icons.Code className="mr-1.5 h-4 w-4" />
+                                Code
                             </TabsTrigger>
-                        </ChatPanelDropdown>
-                        <ChatControls />
+                        </div>
+                        {selectedTab === EditorTabValue.CHAT && <ChatControls />}
                     </TabsList>
                     <ChatHistory isOpen={isChatHistoryOpen} onOpenChange={setIsChatHistoryOpen} />
                     <TabsContent className='h-full overflow-y-auto' value={EditorTabValue.CHAT}>
                         <ChatTab />
+                    </TabsContent>
+                    <TabsContent className='h-full overflow-y-auto' value={EditorTabValue.DEV}>
+                        {/* <DevTab /> */}
+                        hi
                     </TabsContent>
                 </Tabs>
             </ResizablePanel >
